+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog], and this project adheres to
 
 - Add `DurationBuilder.WithMinimum()` and `WithMaximum()`
 - Add `SignedBuilder.WithMinimum()` and `WithMaximum()`
+- Add `UnsignedBuilder.WithMinimum()` and `WithMaximum()`
 
 ## [0.3.1] - 2022-08-24